Refactor move models helper functions to a new module: models_helpers.py
