+++ concent_api/core/subtask_helpers.py
@@ -1,17 +1,12 @@
 from base64 import b64encode
 from logging import getLogger
-from psycopg2 import errorcodes as pg_errorcodes
-from typing import Any, Type
 from typing import List
-from typing import Optional
 from typing import Union
 
 from django.conf import settings
-from django.db import transaction, IntegrityError
-from django.db.models import Model
+from django.db import transaction
 from django.db.models import Q
-from django.db.models import QuerySet
-from django.db.models.base import ModelBase
+
 from golem_messages.message import Message
 from golem_messages.message.concents import ForcePayment
 from golem_messages.message.tasks import SubtaskResultsAccepted
@@ -25,6 +20,7 @@ from core.exceptions import UnsupportedProtocolVersion
 from core.models import DepositClaim
 from core.models import PendingResponse
 from core.models import Subtask
+from core.model_helpers import get_one_or_none
 from core.payments import bankster
 from core.transfer_operations import store_pending_message
 from core.transfer_operations import verify_file_status
@@ -202,7 +198,7 @@ def pre_process_message_related_subtasks(
     for subtask_id in subtask_ids_list:
         with transaction.atomic(using='control'):
             subtask = get_one_or_none(
-                subtask_or_query_set=Subtask.objects.select_for_update(),
+                model_or_query_set=Subtask.objects.select_for_update(),
                 subtask_id=subtask_id,
             )
             if subtask is None:
@@ -294,20 +290,6 @@ def are_subtask_results_accepted_messages_signed_by_the_same_requestor(
     return are_all_signed_by_requestor
 
 
-def get_one_or_none(
-    subtask_or_query_set: Union[Model, QuerySet],
-    **conditions: Any
-)-> Optional[Model]:
-    if isinstance(subtask_or_query_set, ModelBase):
-        instances = subtask_or_query_set.objects.filter(**conditions)
-        assert len(instances) <= 1
-        return None if len(instances) == 0 else instances[0]
-    else:
-        instances = subtask_or_query_set.filter(**conditions)
-        assert len(instances) <= 1
-        return None if len(instances) == 0 else instances[0]
-
-
 def finalize_deposit_claim(
     subtask_id: str,
     concent_use_case: ConcentUseCase,
@@ -345,18 +327,3 @@ def is_state_transition_possible(
     from_: Subtask.SubtaskState,
 ) -> bool:
     return from_ in Subtask.POSSIBLE_TRANSITIONS_TO[to_]  # type: ignore
-
-
-def get_or_create_safely(model: Type[Model], **kwargs: Any) -> Model:
-    try:
-        model_object = model.objects.get_or_create_full_clean(
-            **kwargs,
-        )
-    except IntegrityError as exception:
-        if exception.pgcode == pg_errorcodes.UNIQUE_VIOLATION:
-            model_object = model.objects.get_or_create_full_clean(
-                **kwargs,
-            )
-        else:
-            raise
-    return model_object
